backend/requirements.txt
mysql-connector-python
fastapi[all]
uvicorn
python-multipart
